Finalizando o printf, arrumando erro do hex recebendo um void * e agora passando direto um uunsigned long sem cast, arrumando makefile também e reorganizando as pastas

+++ src/ft_print_hex.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-int	ft_print_hex(unsigned int n, int fd, int uppercase)
+int	ft_print_hex(unsigned long n, int fd, int uppercase)
 {
 	int		count;
 	char	*base;
